# apps/schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BotConnector Control API",
    "version": "1.0.0",
    "description": "Phase 3: Project / Artifact Control API"
  },
  "paths": {
...
    "/api/v1/projects/{projectId}/artifacts": {
      "post": {
        "operationId": "createArtifact",
        "summary": "Create an artifact",
        "tags": [
          "Artifact"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArtifactRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Artifact created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict"
          },
          "422": {
            "description": "Validation error"
          }
        }
      },
      "get": {
        "operationId": "listArtifacts",
        "summary": "List artifacts in a project",
        "tags": [
          "Artifact"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of artifacts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{artifactId}": {
      "get": {
        "operationId": "getArtifact",
        "summary": "Get an artifact",
        "tags": [
          "Artifact"
        ],
        "parameters": [
          {
            "name": "artifactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artifact details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "operationId": "updateArtifact",
        "summary": "Update an artifact",
        "tags": [
          "Artifact"
        ],
        "parameters": [
          {
            "name": "artifactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArtifactRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Artifact updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Revision conflict"
          }
        }
      }
    },
    "/api/v1/artifacts/{artifactId}/versions": {
      "post": {
        "operationId": "createArtifactVersion",
        "summary": "Create an artifact version",
        "tags": [
          "ArtifactVersion"
        ],
        "parameters": [
          {
            "name": "artifactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArtifactVersionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Artifact version created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactVersionResponse"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict"
          }
        }
      },
      "get": {
        "operationId": "listArtifactVersions",
        "summary": "List artifact versions",
        "tags": [
          "ArtifactVersion"
        ],
        "parameters": [
          {
            "name": "artifactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of artifact versions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactVersionListResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtifactManifest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "const": 1
          },
          "id": {
            "type": "string",
            "minLength": 1
          },
          "project_id": {
            "type": "string",
            "minLength": 1
          },
          "type": {
            "type": "string",
            "enum": [
              "web",
              "mobile",
              "design",
              "presentation",
              "document",
              "spreadsheet",
              "visualization",
              "animation"
            ]
          },
          "lifecycle": {
            "type": "string",
            "enum": [
              "draft",
              "generating",
              "valid",
              "ready",
              "published",
              "archived"
            ]
          },
          "revision": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "base_revision": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "current_version_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        },
        "required": [
          "version",
          "id",
          "project_id",
          "type",
          "lifecycle",
          "revision",
          "base_revision",
          "current_version_id",
          "created_at",
          "updated_at"
        ],
        "additionalProperties": false
      },
      "ArtifactVersion": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "const": 1
          },
          "id": {
            "type": "string",
            "minLength": 1
          },
          "artifact_id": {
            "type": "string",
            "minLength": 1
          },
          "sequence": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "source_revision": {
            "type": "string",
            "minLength": 1
          },
          "content_hash": {
            "type": "string",
            "minLength": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        },
        "required": [
          "version",
          "id",
          "artifact_id",
          "sequence",
          "source_revision",
          "content_hash",
          "created_at"
        ],
        "additionalProperties": false
      },
...
      "ApiMetadata": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "const": 1
          },
          "request_id": {
            "type": "string",
            "minLength": 1
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        },
        "required": [
          "version",
          "request_id",
          "timestamp"
        ],
        "additionalProperties": false
      },
...
      "CreateArtifactRequest": {
        "type": "object",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/ArtifactType"
          }
        }
      },
      "UpdateArtifactRequest": {
        "type": "object",
        "required": [
          "revision"
        ],
        "properties": {
          "lifecycle": {
            "$ref": "#/components/schemas/ArtifactLifecycle"
          },
          "revision": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "ArtifactResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ArtifactManifest"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMetadata"
          }
        }
      },
      "ArtifactListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtifactManifest"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMetadata"
          }
        }
      },
      "CreateArtifactVersionRequest": {
        "type": "object",
        "required": [
          "source_revision",
          "content_hash"
        ],
        "properties": {
          "source_revision": {
            "type": "string",
            "minLength": 1
          },
          "content_hash": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "ArtifactVersionResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ArtifactVersion"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMetadata"
          }
        }
      },
      "ArtifactVersionListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtifactVersion"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMetadata"
          }
        }
      },
...
    }
  }
}